+++ src/mol-plugin/behavior/dynamic/custom-props.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2020 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -14,3 +14,5 @@ export { ValenceModel } from './custom-props/computed/valence-model';
 export { SIFTSMapping as BestDatabaseSequenceMapping } from './custom-props/sequence/sifts-mapping';
 
 export { CrossLinkRestraint } from './custom-props/integrative/cross-link-restraint';
+
+export { Streamlines } from './custom-props/volume/streamlines';